This will make Travis sad, but force a remote update on start and status
For start because if I "task start feature/blah" which someone just created,
I want to get their existing branch, not create a new one, even if I haven't
fetched since they created it.

For status because I always want to see the latest status of the branch on
the server if someone else may have deployed it.